+++ examples/c/43_lerp_trait/src/main.c
@@ -1,13 +1,13 @@
 #include <lerp_trait.h>
 
 typedef struct Position {
-    float x;
-    float y;
+    double x;
+    double y;
 } Position;
 
 typedef struct Lerp {
-    float value;
-    float lerp_time;
+    double value;
+    double lerp_time;
 } Lerp;
 
 void DoLerp(ecs_iter_t *it) {
@@ -17,29 +17,29 @@ void DoLerp(ecs_iter_t *it) {
     /* This is the actual component value. Because the system can match any
      * component type, the size is unknown. */
     size_t size = ecs_column_size(it, 2);
-    float *cur = ecs_column_w_size(it, size, 2);    
+    double *cur = ecs_column_w_size(it, size, 2);    
 
     /* These are the trait columns for LerpStart and LerpStop. Because these are
      * trait tags, the system does not know their types at compile time. */
-    float *start = ecs_column_w_size(it, size, 3);
-    float *stop = ecs_column_w_size(it, size, 4);
+    double *start = ecs_column_w_size(it, size, 3);
+    double *stop = ecs_column_w_size(it, size, 4);
 
     /* Apply the lerp. Because we don't know the type of the component we'll
-     * assume that the component consists out of only float values. We can then
-     * iterate the component pointer with increments of sizeof(float) and apply
+     * assume that the component consists out of only double values. We can then
+     * iterate the component pointer with increments of sizeof(double) and apply
      * the lerp to each pointer.
      *
      * A more complete implementation of a lerp could add support for multiple
      * datatypes like double, which could be done by either adding a field to
      * the Lerp trait, or by introducing a Lerp4 and Lerp8 trait. */
     for (int32_t i = 0; i < it->count; i ++) {
-        float lerp = l[i].value + it->delta_time / l[i].lerp_time;
-        bool lerp_done = lerp >= 1.0f;
-        lerp = (lerp * !lerp_done) + (1.0f * lerp_done);
+        double lerp = l[i].value + (double)it->delta_time / l[i].lerp_time;
+        bool lerp_done = lerp >= 1.0;
+        lerp = (lerp * !lerp_done) + (1.0 * lerp_done);
 
         /* Do the actual lerp */
-        for (size_t s = 0; s < size; s += sizeof(float)) {
-            *cur = *start * (1.0f - lerp) + *stop * lerp;
+        for (size_t s = 0; s < size; s += sizeof(double)) {
+            *cur = *start * (1.0 - lerp) + *stop * lerp;
 
             cur ++;
             start ++;
